for merge

+++ include/Rasterizer.hpp
@@ -17,7 +17,7 @@ private:
     void draw_triangle(Vertex v1, Vertex v2, Vertex v3, Mat4 &transfo);
     void draw_point(Vertex v1, Mat4 &transfo);
     void draw_sphere();
-    void set_pixel_color(uint x, uint y, uint z, const Color &c);
+    void set_pixel_color(uint x, uint y, float z, const Color &c);
     void upload_texture() const;
 
 public:
